+++ CHANGES.md
@@ -7,11 +7,11 @@ Change History
 
 New features:
 
-- [Pull #17](https://github.com/cpburnz/minecraft-prometheus-exporter/pull/17): Add entity tracking with the "mc_entity_list{type, id}" metric. This can be disabled by setting "collector.mc_entity_list" to "false" in the config.
+- [Pull #17](https://github.com/cpburnz/minecraft-prometheus-exporter/pull/17): Add entity tracking with the "mc_entities_total{dim, dim_id, id, type}" metric. This can be disabled by setting "collector.mc_entities" to "false" in the config.
 
 Bug fixes:
 
-- The "mc_player_list" metric sets the "id" attribute with the profile UUID when it is available.
+- The "mc_player_list" metric sets the "id" label with the profile UUID only when it is available.
 
 
 1.0.0 for MC 1.7.10 (2023-12-17)